PAKTON Framework/Archivist/requirements.txt
langchain-core==0.3.41
langchain-community==0.3.19
langchain==0.3.20
pydantic==2.10.6
langgraph==0.3.5
langgraph-checkpoint==2.0.26
langgraph-checkpoint-postgres==2.0.21
psycopg[binary]>=3.2.12
langsmith==0.3.11
langchain-openai==0.3.7
notebook==7.3.2
setuptools
pinecone==5.4.2
langchain-pinecone==0.2.3
pypdf==5.9.0
python-docx==1.2.0
langchain-chroma==0.2.2
unstructured==0.18.20
langchain-unstructured==0.1.6
langchain-aws==0.2.15
docx2txt==0.9

# Local package dependencies
# Interrogator will be installed separately via Dockerfile